requirements.txt
Django>=3.0.0,<3.1.2
djangorestframework>=3.0.0,<3.11.2
Pillow>=5.3.0,<7.2.0
